feat: added statistics_workload to nemesis tool (#24514)
