+++ ydb/tests/stability/tool/__main__.py
@@ -111,6 +111,14 @@ DICT_OF_PROCESSES = {
                 echo "Stopped"
             fi"""
     },
+    'statistics' : {
+        'status' : """
+            if ps aux | grep -E "/Berkanavt/nemesis/bin/statistics|/tmp/statistics_wrapper.sh" | grep -v grep > /dev/null; then
+                echo "Running"
+            else
+                echo "Stopped"
+            fi"""
+    },
     'workload_log_column' : {
         'status' : """
             if ps aux | grep -E "/Berkanavt/nemesis/bin/ydb_cli.*workload.*log.*run.*bulk_upsert.*log_workload_column|/tmp/workload_log_column_wrapper.sh" | grep -v grep > /dev/null; then
@@ -218,6 +226,7 @@ class CustomArgumentParser(argparse.ArgumentParser):
                 "start_workload_log_column": "Start log workload with column storage",
                 "start_workload_log_row": "Start log workload with row storage",
                 "start_workload_topic": "Start topic workload",
+                "start_statistics_workload": "Start statistics workload",
                 "stop_workloads": "Stop all workloads",
                 "stop_workload": "Stop a specific workload (requires --name)",
                 "perform_checks": "Run safety and liveness checks on the cluster",
@@ -478,7 +487,8 @@ class StabilityCluster:
 
         # На всякий случай убиваем все потенциальные процессы рабочих нагрузок
         node.ssh_command(
-            'pkill -f "SCREEN.*workload\\|simple_queue\\|olap_workload\\|oltp_workload\\|node_broker_workload\\|transfer_workload\\|s3_backups_workload"',
+            'pkill -f "SCREEN.*workload\\|simple_queue\\|olap_workload\\|oltp_workload\\|node_broker_workload\\|transfer_workload\\|s3_backups_workload\\|\
+                statistics_workload\\|kafka_workload\\|topic_kafka_workload\\|ctas_workload\\|topic_workload\\|"',
             raise_on_error=False
         )
 
@@ -1073,6 +1083,31 @@ handle_timeout() {{
     echo "[$(date +'%Y-%m-%d %H:%M:%S.%N')] Performing s3_backups_workload specific cleanup"
     # Убиваем процессы s3_backups_workload
     pkill -9 -f "s3_backups_workload" || true
+
+  elif [[ "{base_command}" == *"topic_workload"* ]]; then
+    echo "[$(date +'%Y-%m-%d %H:%M:%S.%N')] Performing topic_workload specific cleanup"
+    # Убиваем процессы topic_workload
+    pkill -9 -f "topic_workload" || true
+
+  elif [[ "{base_command}" == *"ctas_workload"* ]]; then
+    echo "[$(date +'%Y-%m-%d %H:%M:%S.%N')] Performing ctas_workload specific cleanup"
+    # Убиваем процессы ctas_workload
+    pkill -9 -f "ctas_workload" || true
+
+  elif [[ "{base_command}" == *"kafka_workload"* ]]; then
+    echo "[$(date +'%Y-%m-%d %H:%M:%S.%N')] Performing kafka_workload specific cleanup"
+    # Убиваем процессы kafka_workload
+    pkill -9 -f "kafka_workload" || true
+
+  elif [[ "{base_command}" == *"statistics_workload"* ]]; then
+    echo "[$(date +'%Y-%m-%d %H:%M:%S.%N')] Performing statistics_workload specific cleanup"
+    # Убиваем процессы statistics_workload
+    pkill -9 -f "statistics_workload" || true
+
+  else [[ "{base_command}" == *"topic_kafka_workload"* ]]; then
+    echo "[$(date +'%Y-%m-%d %H:%M:%S.%N')] Performing topic_kafka_workload specific cleanup"
+    # Убиваем процессы topic_kafka_workload
+    pkill -9 -f "topic_kafka_workload" || true
   fi
 
   echo "[$(date +'%Y-%m-%d %H:%M:%S.%N')] Emergency cleanup completed"
@@ -1289,6 +1324,7 @@ Common usage scenarios:
         "start_workload_log_column": "Start log workload with column storage",
         "start_workload_log_row": "Start log workload with row storage",
         "start_workload_topic": "Start topic workload",
+        "start_statistics_workload": "Start statistics workload",
         "stop_workloads": "Stop all workloads",
         "stop_workload": "Stop a specific workload (requires --name)",
         "perform_checks": "Run safety and liveness checks on the cluster",
@@ -1323,6 +1359,7 @@ Common usage scenarios:
             "start_ctas_workload",
             "start_workload_log", "start_workload_log_column", "start_workload_log_row",
             "start_workload_topic",
+            "start_statistics_workload",
         ]
     }
 
@@ -1635,6 +1672,14 @@ def main():
                     node,
                     'ctas_workload',
                     f'/Berkanavt/nemesis/bin/ctas_workload --database /Root/db1 --path {node_id}'
+                )
+            stability_cluster.get_state()
+        if action == "start_statistics_workload":
+            for node_id, node in enumerate(stability_cluster.kikimr_cluster.nodes.values()):
+                stability_cluster._clean_and_start_workload(
+                    node,
+                    'statistics_workload',
+                    f'/Berkanavt/nemesis/bin/statistics_workload --database /Root/db1 --prefix statistics_workload/{node_id}/table'
                 )
             stability_cluster.get_state()
         if action == "stop_workloads":
