+++ ydb/tests/stress/statistics_workload/workload/ya.make
@@ -1,0 +1,15 @@
+PY3_LIBRARY()
+
+PY_SRCS(
+    __init__.py
+)
+
+PEERDIR(
+    ydb/tests/library
+    ydb/public/sdk/python
+    library/python/monlib
+    ydb/public/sdk/python/enable_v3_new_behavior
+)
+
+
+END()
